requirements/local.txt
Django==1.11.3
django-registration==2.2
pytz==2017.2
